Resolve issue 2 and 6
Make texts on the website unselectable.

Add 9 colors to the shape.

When drawing a arc, if the startAngle and endAngle are the same,
it will draw a ring instead of a dot.

When the shape was chosen, show its frame. If user click the
same shape continuously, the frame will disappeared, and the
shape will be sent to the lowest order.